+++ kuma/api/v1/plus.py
@@ -7,7 +7,7 @@ from django.shortcuts import get_object_or_404
 from django.views.decorators.http import require_http_methods
 from ratelimit.decorators import ratelimit
 
-from kuma.mdnplusplus.models import LandingPageSurvey
+from kuma.plus.models import LandingPageSurvey
 
 
 @ratelimit(key="user_or_ip", rate="100/m", block=True)
