+++ kuma/api/v1/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path, re_path
 
-from . import mdnplusplus, search, views
+from . import plus, search, views
 
 urlpatterns = [
     re_path(r"^whoami/?$", views.whoami, name="api.v1.whoami"),
@@ -13,9 +13,9 @@ urlpatterns = [
         name="api.v1.send_subscriptions_feedback",
     ),
     path(
-        "mdnplusplus/landing-page-survey/",
-        mdnplusplus.landing_page_survey,
-        name="api.v1.mdnplusplus.landing_page_survey",
+        "plus/landing-page-survey/",
+        plus.landing_page_survey,
+        name="api.v1.plus.landing_page_survey",
     ),
     path("subscriptions/", views.subscriptions, name="api.v1.subscriptions"),
     path("stripe_hooks/", views.stripe_hooks, name="api.v1.stripe_hooks"),
