--- kuma/mdnplusplus/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class MDNPlusplusConfig(AppConfig):
-    name = "kuma.mdnplusplus"
-    verbose_name = "MDN++"
