+++ kuma/plus/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PlusConfig(AppConfig):
+    name = "kuma.plus"
+    verbose_name = "Plus"
